requirements.txt
Flask
gunicorn
python-dotenv
pymysql
google-generativeai
tiktoken
pydantic-settings
pypdf
langchain 
langchain-core 
langchain-google-genai 
psycopg2-binary
werkzeug
blinker
redis
itsdangerous>=2.2.0
jinja2>=3.1.2
google-api-python-client
pydantic
tqdm
langchain-text-splitters
langsmith
PyYAML
SQLAlchemy

# Core dependencies
annotated-types>=0.6.0
pydantic-core>=2.33.0
typing-inspection>=0.4.0

# Google API dependencies
google-auth-httplib2>=0.2.0,<1.0.0
httplib2>=0.19.0,<1.0.0
uritemplate>=3.0.1,<5

# Langsmith dependencies  
httpx>=0.23.0,<1
orjson>=3.9.14
requests-toolbelt>=1.0.0
zstandard>=0.23.0

# Langchain dependencies
jsonpatch>=1.33,<2.0
tenacity>=8.1.0,<10.0.0,!=8.4.0

# Fixed dependency conflicts
google-ai-generativelanguage>=0.6.15
filetype>=1.2,<2
regex>=2022.1.18
alembic
pip-tools
numpy
python-docx
backoff
flask-jwt-extended
bcrypt